+++ modules/analysis/rosbag-image-pipeline/image_dags/ros_image_pipeline.py
@@ -18,15 +18,13 @@ from airflow import DAG, settings
 from airflow.contrib.hooks.aws_hook import AwsHook
 from airflow.exceptions import AirflowException
 from airflow.models import Connection
-from airflow.operators.python import PythonOperator
-from airflow.providers.amazon.aws.operators.batch import AwsBatchOperator
-from airflow.providers.amazon.aws.operators.emr_containers import EMRContainerOperator
-from airflow.providers.amazon.aws.sensors.emr_containers import EMRContainerSensor
+from airflow.operators.python import PythonOperator, get_current_context, task
+from airflow.providers.amazon.aws.operators.batch import BatchOperator
+from airflow.providers.amazon.aws.operators.emr import EmrServerlessStartJobOperator
+from airflow.providers.amazon.aws.sensors.emr import EmrServerlessJobSensor
 from airflow.utils.dates import days_ago
 from airflow.utils.task_group import TaskGroup
 from boto3.dynamodb.conditions import Key
-from emr_serverless.operators.emr import EmrServerlessStartJobOperator
-from emr_serverless.sensors.emr import EmrServerlessJobSensor
 from sagemaker.network import NetworkConfig
 from sagemaker.processing import ProcessingInput, ProcessingOutput, Processor
 
@@ -199,12 +197,16 @@ def get_job_name(suffix="") -> str:
 
 
 def png_batch_operation(**kwargs):
+
+    logger.info(f"kwargs at png_batch_operations is {kwargs}")
+
     ti = kwargs["ti"]
     ds = kwargs["ds"]
     array_size = ti.xcom_pull(task_ids="create-batch-of-drives", key="return_value")
     batch_id = kwargs["dag_run"].run_id
+    context = get_current_context()
 
-    op = AwsBatchOperator(
+    op = BatchOperator(
         task_id="submit_batch_job_op",
         job_name=get_job_name("png"),
         job_queue=ON_DEMAND_JOB_QUEUE_ARN,
@@ -223,7 +225,8 @@ def png_batch_operation(**kwargs):
         },
     )
 
-    op.execute(ds)
+    # op.execute(ds)
+    op.execute(context)
 
 
 def parquet_operation(**kwargs):
@@ -231,8 +234,9 @@ def parquet_operation(**kwargs):
     ds = kwargs["ds"]
     array_size = ti.xcom_pull(task_ids="create-batch-of-drives", key="return_value")
     batch_id = kwargs["dag_run"].run_id
+    context = get_current_context()
 
-    op = AwsBatchOperator(
+    op = BatchOperator(
         task_id="submit_parquet_job_op",
         job_name=get_job_name("parq"),
         job_queue=FARGATE_JOB_QUEUE_ARN,
@@ -249,7 +253,7 @@ def parquet_operation(**kwargs):
         },
     )
 
-    op.execute(ds)
+    op.execute(context)
 
 
 def sagemaker_yolo_operation(**kwargs):